Remove logging and verbosity
There is currently nothing to log or additional output to display. It is
just getting in the way for the time being. Maybe I will add it back in
at a later time when it is needed. I am just not sure how the logging
will work with the STDIO streams.