Change commands JSON format
A version field is included to allow upgrades and file format changes in
the future if needed.

The `-c` flag is added, which outputs the commands JSON file format
instead of the asciicast file format. This will allow creating commands
JSON format and eventually "chaining" multiple commands through
redirection and pipes.