Add interactive mode
If nothing is piped or redirected to stdin and no <INPUT> and <OUTPUT>
arguments are specified, the application enters an simple interactive
mode where commands can be typed within the terminal, send to a child
shell, the output brought back, and a cast file is created. This is a
very simple and naive implementation. Basically, lines are read from
stdin and the output from each command is written to stdout of the
parent process (this application) while also be added as the output for
the command. When EOF is reached, i.e. CTRL+D, then the interactive
session is complete and the asciicast file is printed. This works best
if the `-o,--output` option is used.