Change CLI for input and output
If no `<outputs>` are provided, then the `<input>` is treated as a shell
command to be executed and the execution outputs to be used. This is
similar to the interactive mode but for a single input. If no input is
provided, then the interactive mode is entered. To avoid executing the
input, i.e. a literal, then an empty "" string should be used as the
second positional argument.