Add command line parsing to interactive mode
This is a little more involved than originally planned. More work is
needed to separate out arguments that are used in interactive mode
versus the CLI and vice-versa, but it is the first working
implementation.

The `Timings` struct should be used as a structopt type and flattened
into the Cli struct, too.

Closes #21.