Change to use shell option for executing commands
This is related to #27. It does NOT resolve the `-c` argument usage
being customizable, i.e. Windows console, but it will change the shell
used to execute the comand based on the `SHELL` environment variable and
then overwritten by the `--shell` option. The `--shell` option is
available in the CLI and interactive mode.