Add ability to change shell and shell argument
Instead of adding "magic" as discussed in #27, a Windows user will
simply have to explicitly define the shell and execute string flag in an
environment variable or command line option. The `--execute-string-flag`
option has been added to override the UNIX shell default and the
possible values are limited to an enum instead of just passing the
string straight to the command, which feels like a security problem
waiting to happen.

This does change the CLI slightly. The `-E` short argument is used for
the `--execute-string-flag` option and the end and begin delay short
flags are changed to lowercase. This makes arguments more consistent
since the `--shell` option uses the uppercase `-S` short flag.

Closes #27.