Refactor to more modules
The "app" components are moved to their own `app` module. This makes the
`main.rs` very short and sweet, which leaves plenty of room for writing
application-level documentation in the form of module-level doc comments
to be added without making the file too large.

This also allows the visibility of some modules to be changed to
private, which highlighted some functions, methods, and constants were
unused. This is related to #6 and #7.