Change to reading into buffer
There are conditions when stdin is not a TTY, i.e. piped or redirected,
but empty. An empty stdin should indicate to use a new sequence as
opposed to appending to an existing sequence, but if stdin is NOT a TTY
and no input file or input arguments are specified, then the interactive
mode should be initiated. The EOF with an empty stdin occurs in testing
but an error will always occur because the JSON deserializers do not
like empty strings.

The solution was to read stdin into a buffer until the EOF is reached
and check if the buffer is empty _before_ deserializing. This occurs
after the check for TTY, so interactive mode can still be entered.